requirements.txt
streamlit
openai
python-dotenv
pyautogen
requests
beautifulsoup4
